pyproject.toml
[tool.poetry]
name = "json-serialize"
version = "0.1.0"
description = "Examples for the blog about JSON serialize"
authors = ["Pedro López Mareque <pedro.lopez.mareque@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.0.1"
pytest-approvaltests = "^0.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
